requirements.txt
numpy
pandas
matplotlib
scipy
streamlit
seaborn
plotly.express 